requirements.txt
google-cloud-speech
google-api-python-client==1.6.4